os/drivers/wireless/relatek: Modify rtk api header directory
- Applying header location and refactoring
- Clean up entire code
