+++ external/Kconfig
@@ -21,7 +21,6 @@ source "$EXTERNALDIR/libopus/Kconfig"
 source "$EXTERNALDIR/stress_tool/Kconfig"
 source "$EXTERNALDIR/swepd/Kconfig"
 source "$EXTERNALDIR/esp_idf_port/Kconfig"
-source "$EXTERNALDIR/rtk_wifi/Kconfig"
 source "$EXTERNALDIR/netutils/Kconfig.netutil"
 source "$EXTERNALDIR/dhcpd/Kconfig.protocol"
 source "$EXTERNALDIR/dhcpc/Kconfig.protocol"
