--- external/rtk_wifi/Kconfig
@@ -1,12 +1,0 @@
-#
-# For a description of the syntax of this configuration file,
-# see kconfig-language at https://www.kernel.org/doc/Documentation/kbuild/kconfig-language.txt
-#
-config RTK_WIFI
-	bool "REALTEK Wi-Fi API"
-	default n
-	select SELECT_RTK_WLAN
-	depends on !LWNL80211_RTK
-	---help---
-		Enable support for the REALTEK WI-Fi Networking API.
-
