+++ os/drivers/wireless/realtek/Kconfig
@@ -1,7 +1,7 @@
 menuconfig RTK_WLAN
 	bool "Realtek Wireless Module support"
 	default n
-	
+
