+++ os/drivers/wireless/realtek/wifi_util_interface/wifi_conf.c
@@ -30,7 +30,6 @@
 
 #include "usb_io_realtek.h"
 #include "rtk_lwip_netconf.h"
-#include "rtk_wifi_utils.h"
 #include <debug.h>
 
 /******************************************************
@@ -55,6 +54,7 @@
 /******************************************************
  *                 Type Definitions
  ******************************************************/
+
 
 /******************************************************
  *               Variables Declarations
