+++ os/include/tinyara/wifi/rtk/wifi_constants.h
@@ -148,7 +148,7 @@ enum {
 typedef unsigned long rtw_security_t;
 
 enum {
-    RTW_ENCRYPTION_UNKNOWN = 0,                 
+    RTW_ENCRYPTION_UNKNOWN = 0,
     RTW_ENCRYPTION_OPEN = 1,
     RTW_ENCRYPTION_WEP40 = 2,
     RTW_ENCRYPTION_WPA_TKIP = 3,
@@ -157,7 +157,7 @@ enum {
     RTW_ENCRYPTION_WPA2_AES = 6,
     RTW_ENCRYPTION_WPA2_MIXED = 7,
     RTW_ENCRYPTION_WEP104 = 9,
-    RTW_ENCRYPTION_UNDEF = 0xFF,                 
+    RTW_ENCRYPTION_UNDEF = 0xFF,
 };
 typedef unsigned long rtw_encryption_t;
 
@@ -373,7 +373,7 @@ enum {
 	RTW_COUNTRY_YT,
 	RTW_COUNTRY_ZA,
 	RTW_COUNTRY_ZW,
-	
+
 	RTW_COUNTRY_MAX
 
 };
@@ -552,7 +552,7 @@ typedef unsigned long rtw_tx_pwr_percentage_t;
 enum _WIFI_EVENT_INDICATE{
 	WIFI_EVENT_CONNECT = 0,
 	WIFI_EVENT_DISCONNECT = 1,
-	WIFI_EVENT_FOURWAY_HANDSHAKE_DONE = 2,	
+	WIFI_EVENT_FOURWAY_HANDSHAKE_DONE = 2,
 	WIFI_EVENT_SCAN_RESULT_REPORT = 3,
 	WIFI_EVENT_SCAN_DONE = 4,
 	WIFI_EVENT_RECONNECTION_FAIL = 5,
